Start introducing a map reduce to parallelize test runs
We split the test suite in 10 parts by abusing travis' matrix
feature. Reduce is done by hand, by analyzing the dashboard of travis,
any failing test being critical.

Because nosetsts --processes just doesn't work with youtube-dl test
suite, we route around it by first enumerating tests. This is a bit long
because nosetests needs to find and run all tests files in order to
enumerate them, but it should be at most 30 seconds, while the test
suite can take more than 2 hours on travis' infrastructure.

By doing this, we ensure we'll be able to run the tests faster, since
they are mostly I/O (network) bound due to the nature of the project.